Update codes for RL78/G23 Fast Prototyping Board.

+++ source/include/core_mqtt_serializer.h
@@ -256,7 +256,7 @@ typedef struct MQTTPacketInfo
     /**
      * @brief Length of remaining serialized data.
      */
-    size_t remainingLength;
+    uint32_t remainingLength;
 } MQTTPacketInfo_t;
 
 /**
